chore: refactor examples.
- move `examples/basic` to `examples/full`
- remove `examples/ssr-spa`
