+++ .github/workflows/build.yml
@@ -12,8 +12,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
         example:
-          - examples/basic
-          - examples/ssr-spa
+          - examples/full
 
     steps:
       - uses: actions/checkout@v4
