+++ examples/full/layouts/style.css
