+++ examples/full/package.json
