+++ examples/full/pages/+Head.tsx
