+++ examples/full/pages/+Layout.tsx
@@ -1,9 +1,9 @@
-import "./style.css";
+import "../layouts/style.css";
 import logoUrl from "../assets/logo.svg";
-import { Link } from "../components/Link.js";
+import { Link } from "../components/Link";
 import type { JSX } from "solid-js";
 
-export default function LayoutDefault(props: { children?: JSX.Element }) {
+export function Layout(props: { children?: JSX.Element }) {
   return (
     <div
       style={{
@@ -14,8 +14,11 @@ export default function LayoutDefault(props: { children?: JSX.Element }) {
     >
       <Sidebar>
         <Logo />
-        <Link href="/">SSR</Link>
-        <Link href="/spa">SPA</Link>
+        <Link href="/">Welcome</Link>
+        <Link href="/star-wars">Data Fetching</Link>
+        <Link href="/without-ssr">Without SSR</Link>
+        <Link href="/starship">Nested Layout 1</Link>
+        <Link href="/stardust">Nested Layout 2</Link>
       </Sidebar>
       <Content>{props.children}</Content>
     </div>
