+++ examples/full/pages/+config.ts
