+++ examples/full/pages/_error/+Page.tsx
