+++ examples/full/pages/index/+Page.tsx
