+++ examples/full/pages/index/Counter.tsx
