+++ examples/full/pages/star-wars/@id/+Page.tsx
