+++ examples/full/pages/star-wars/@id/+data.ts
