+++ examples/full/pages/star-wars/@id/+title.ts
