+++ examples/full/pages/star-wars/index/+Page.tsx
