+++ examples/full/pages/star-wars/index/+data.ts
