+++ examples/full/pages/star-wars/index/+title.ts
