+++ examples/full/pages/star-wars/types.ts
