+++ examples/full/pages/stardust/+Layout.tsx
