+++ examples/full/pages/stardust/+Page.tsx
