+++ examples/full/pages/starship/+Layout.tsx
