+++ examples/full/pages/starship/+Page.tsx
