+++ examples/full/pages/starship/spec/+Page.tsx
