+++ examples/full/pages/without-ssr/+Page.tsx
