+++ examples/full/pages/without-ssr/+config.ts
