+++ examples/full/tsconfig.json
