+++ examples/full/vite.config.ts
