--- examples/ssr-spa/README.md
@@ -1,9 +1,0 @@
-Example of selecting SPA or SSR on a per-page basis. See
-[SPA vs SSR (and more)](https://vike.dev/SPA-vs-SSR).
-
-```bash
-git clone git@github.com:vikejs/vike-solid
-cd vike-solid/examples/ssr-spa/
-npm install
-npm run dev
-```
