--- examples/ssr-spa/package.json
@@ -1,19 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test": "tsc --noEmit"
-  },
-  "dependencies": {
-    "solid-js": "^1.8.17",
-    "vike": "^0.4.177",
-    "vike-solid": "link:vike-solid"
-  },
-  "devDependencies": {
-    "typescript": "^5.4.5",
-    "vite": "^5.3.1"
-  },
-  "type": "module"
-}
