--- examples/ssr-spa/pages/Head.tsx
@@ -1,12 +1,0 @@
-// Default <head> (can be overridden by pages)
-
-import logoUrl from "../assets/logo.svg";
-
-export default function Head() {
-  return (
-    <>
-      <meta name="description" content="Demo showcasing Vike + Solid" />
-      <link rel="icon" href={logoUrl} />
-    </>
-  );
-}
