--- examples/ssr-spa/pages/spa/+Page.tsx
@@ -1,11 +1,0 @@
-export default function Page() {
-  // Will be printed only in the browser:
-  console.log("Rendering SPA page");
-
-  return (
-    <>
-      <h1>My vike-solid app</h1>
-      This page is rendered only in the browser.
-    </>
-  );
-}
