--- examples/ssr-spa/pages/spa/+config.ts
@@ -1,5 +1,0 @@
-import type { Config } from "vike/types";
-
-export default {
-  ssr: false, // SPA
-} satisfies Config;
