--- examples/ssr-spa/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "target": "ES2021",
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "types": ["vike-solid/client"],
-    "jsx": "preserve",
-    "jsxImportSource": "solid-js",
-    "skipLibCheck": true,
-    "esModuleInterop": true
-  }
-}
