--- examples/ssr-spa/vite.config.ts
@@ -1,7 +1,0 @@
-import vikeSolid from "vike-solid/vite";
-import vike from "vike/plugin";
-import type { UserConfig } from "vite";
-
-export default {
-  plugins: [vike(), vikeSolid()],
-} satisfies UserConfig;
